Properly determine length of directory listing html
Python is using utf-8 strings by default and the len of the string is
the number of characters instead of the number of bytes of the encoded
string. This leads to an incorrect Content-Length header if there are
any non-ascii characters in the returned filenames.

Hardcoding utf-8 since there are a bunch of other places that do the
same, e.g. WebDAVServer#L102, WebDAVServer#L250.